+++ src/util_types/mutator_set/root_and_paths.rs
@@ -1,16 +1,4 @@
-use std::collections::hash_map::Entry;
-use std::collections::HashMap;
-
-use itertools::Itertools;
-use proptest::arbitrary::Arbitrary;
-use proptest::collection::vec;
-use proptest::strategy::BoxedStrategy;
-use proptest::strategy::Just;
-use proptest::strategy::Strategy;
-use proptest_arbitrary_interop::arb;
 use tasm_lib::prelude::Digest;
-
-use crate::models::blockchain::shared::Hash;
 
 #[derive(Debug, Clone, Default)]
 pub struct RootAndPaths {
@@ -18,109 +6,129 @@ pub struct RootAndPaths {
     pub paths: Vec<Vec<Digest>>,
 }
 
-impl Arbitrary for RootAndPaths {
-    type Parameters = (usize, Vec<(u64, Digest)>);
-    type Strategy = BoxedStrategy<Self>;
+#[cfg(any(test, feature = "arbitrary-impls"))]
+pub mod neptune_arbitrary {
+    use std::collections::hash_map::Entry;
+    use std::collections::HashMap;
 
-    fn arbitrary_with(parameters: Self::Parameters) -> Self::Strategy {
-        let (tree_height_proper, indices_and_leafs_proper) = parameters;
-        assert!(
-            indices_and_leafs_proper
-                .iter()
-                .map(|(idx, _)| idx)
-                .all_unique(),
-            "indices are not all unique"
-        );
-        assert!(
-            indices_and_leafs_proper
-                .iter()
-                .all(|(i, _l)| (*i as u128) < 1u128 << tree_height_proper),
-            "some or all indices are too large; don't fit in a tree of height {tree_height_proper}"
-        );
-        let upper_bound_num_digests = tree_height_proper * indices_and_leafs_proper.len() + 1;
+    use itertools::Itertools;
+    use proptest::arbitrary::Arbitrary;
+    use proptest::collection::vec;
+    use proptest::strategy::BoxedStrategy;
+    use proptest::strategy::Just;
+    use proptest::strategy::Strategy;
+    use proptest_arbitrary_interop::arb;
 
-        let tree_height_strategy = Just(tree_height_proper);
-        let indices_and_leafs_strategy = Just(indices_and_leafs_proper);
-        let vec_digest_strategy = vec(arb::<Digest>(), upper_bound_num_digests);
-        (
-            tree_height_strategy,
-            indices_and_leafs_strategy,
-            vec_digest_strategy,
-        )
-            .prop_map(|(tree_height, indices_and_leafs, mut digests)| {
-                assert!(indices_and_leafs.iter().all(|(i, _l)| (*i as u128) < (1u128 << tree_height)), "indices too large for tree of height: {}", tree_height);
-                // populate nodes dictionary with leafs
-                let mut nodes = HashMap::new();
-                for &(index, leaf) in &indices_and_leafs {
-                    let node_index = (index as u128) + (1u128 << tree_height);
-                    nodes.insert(node_index, leaf);
-                }
+    use super::*;
+    use crate::models::blockchain::shared::Hash;
 
-                let by_layer = |index: u128, layer: usize| {
-                    let sub_tree_height = tree_height - layer;
-                    let layer_start = 1u128 << sub_tree_height;
-                    let layer_stop = 1u128 << (sub_tree_height + 1);
-                    index >= layer_start && index < layer_stop
-                };
+    impl Arbitrary for RootAndPaths {
+        type Parameters = (usize, Vec<(u64, Digest)>);
+        type Strategy = BoxedStrategy<Self>;
 
-                // walk up tree layer by layer
-                // when we need nodes not already present, sample at random
-                for layer in 0..tree_height {
-                    let mut working_indices = nodes
-                        .keys()
-                        .copied()
-                        .filter(|&i| by_layer(i, layer))
-                        .collect_vec();
-                    working_indices.sort();
-                    working_indices.dedup();
-                    for wi in working_indices {
-                        let wi_odd = wi | 1;
-                        if let Entry::Vacant(entry) = nodes.entry(wi_odd) {
-                            entry.insert(digests.pop().unwrap());
-                        }
-                        let wi_even = wi_odd ^ 1;
-                        if let Entry::Vacant(entry) = nodes.entry(wi_even) {
-                            entry.insert(digests.pop().unwrap());
-                        }
-                        let hash = Hash::hash_pair(nodes[&wi_even], nodes[&wi_odd]);
-                        nodes.insert(wi >> 1, hash);
-                    }
-                }
-
-                // read out root, even if no paths were traversed
-                let root = *nodes.get(&1).unwrap_or(&digests.pop().unwrap());
-
-                // read out paths
-                let paths = indices_and_leafs
+        fn arbitrary_with(parameters: Self::Parameters) -> Self::Strategy {
+            let (tree_height_proper, indices_and_leafs_proper) = parameters;
+            assert!(
+                indices_and_leafs_proper
                     .iter()
-                    .map(|(leaf_idx, _)| (*leaf_idx as u128) + (1u128 << tree_height))
-                    .map(|node_idx| {
-                        (0..tree_height)
-                            .map(|layer_idx| {
-                                nodes
-                                    .get(&((node_idx >> layer_idx) ^ 1u128))
-                                    .unwrap_or_else(|| {
-                                        panic!(
-                                            "node index \n{} \nnot present in node dictionary!\ndo have indices:\n{}",
-                                            (node_idx >> layer_idx) ^ 1,
-                                            nodes.keys().join("\n")
-                                        )
-                                    })
-                            })
-                            .copied()
-                            .collect_vec()
-                    })
-                    .collect_vec();
+                    .map(|(idx, _)| idx)
+                    .all_unique(),
+                "indices are not all unique"
+            );
+            assert!(
+                indices_and_leafs_proper
+                    .iter()
+                    .all(|(i, _l)| (*i as u128) < 1u128 << tree_height_proper),
+                "some or all indices are too large; don't fit in a tree of height {tree_height_proper}"
+            );
+            let upper_bound_num_digests = tree_height_proper * indices_and_leafs_proper.len() + 1;
 
-                RootAndPaths { root, paths }
-            })
-            .boxed()
+            let tree_height_strategy = Just(tree_height_proper);
+            let indices_and_leafs_strategy = Just(indices_and_leafs_proper);
+            let vec_digest_strategy = vec(arb::<Digest>(), upper_bound_num_digests);
+            (
+                tree_height_strategy,
+                indices_and_leafs_strategy,
+                vec_digest_strategy,
+            )
+                .prop_map(|(tree_height, indices_and_leafs, mut digests)| {
+                    assert!(indices_and_leafs.iter().all(|(i, _l)| (*i as u128) < (1u128 << tree_height)), "indices too large for tree of height: {}", tree_height);
+                    // populate nodes dictionary with leafs
+                    let mut nodes = HashMap::new();
+                    for &(index, leaf) in &indices_and_leafs {
+                        let node_index = (index as u128) + (1u128 << tree_height);
+                        nodes.insert(node_index, leaf);
+                    }
+
+                    let by_layer = |index: u128, layer: usize| {
+                        let sub_tree_height = tree_height - layer;
+                        let layer_start = 1u128 << sub_tree_height;
+                        let layer_stop = 1u128 << (sub_tree_height + 1);
+                        index >= layer_start && index < layer_stop
+                    };
+
+                    // walk up tree layer by layer
+                    // when we need nodes not already present, sample at random
+                    for layer in 0..tree_height {
+                        let mut working_indices = nodes
+                            .keys()
+                            .copied()
+                            .filter(|&i| by_layer(i, layer))
+                            .collect_vec();
+                        working_indices.sort();
+                        working_indices.dedup();
+                        for wi in working_indices {
+                            let wi_odd = wi | 1;
+                            if let Entry::Vacant(entry) = nodes.entry(wi_odd) {
+                                entry.insert(digests.pop().unwrap());
+                            }
+                            let wi_even = wi_odd ^ 1;
+                            if let Entry::Vacant(entry) = nodes.entry(wi_even) {
+                                entry.insert(digests.pop().unwrap());
+                            }
+                            let hash = Hash::hash_pair(nodes[&wi_even], nodes[&wi_odd]);
+                            nodes.insert(wi >> 1, hash);
+                        }
+                    }
+
+                    // read out root, even if no paths were traversed
+                    let root = *nodes.get(&1).unwrap_or(&digests.pop().unwrap());
+
+                    // read out paths
+                    let paths = indices_and_leafs
+                        .iter()
+                        .map(|(leaf_idx, _)| (*leaf_idx as u128) + (1u128 << tree_height))
+                        .map(|node_idx| {
+                            (0..tree_height)
+                                .map(|layer_idx| {
+                                    nodes
+                                        .get(&((node_idx >> layer_idx) ^ 1u128))
+                                        .unwrap_or_else(|| {
+                                            panic!(
+                                                "node index \n{} \nnot present in node dictionary!\ndo have indices:\n{}",
+                                                (node_idx >> layer_idx) ^ 1,
+                                                nodes.keys().join("\n")
+                                            )
+                                        })
+                                })
+                                .copied()
+                                .collect_vec()
+                        })
+                        .collect_vec();
+
+                    RootAndPaths { root, paths }
+                })
+                .boxed()
+        }
     }
 }
 
 #[cfg(test)]
 mod test {
+    use itertools::Itertools;
+    use proptest::collection::vec;
     use proptest::prelude::*;
+    use proptest_arbitrary_interop::arb;
     use tasm_lib::twenty_first::util_types::merkle_tree::MerkleTreeInclusionProof;
     use test_strategy::proptest;
